AudiobookMaker/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
aiofiles==23.2.1
openai==1.12.0
requests==2.31.0
pydub==0.25.1
ebooklib==0.18
beautifulsoup4==4.12.2
PyPDF2==3.0.1
python-dotenv==1.0.0
httpx==0.26.0
numpy==1.26.3
